+++ pysrc/constants.py
@@ -4,7 +4,9 @@ import numpy as np
 halfpi = 0.5 * np.pi
 twopi = 2.0 * np.pi
 lzero = np.log(1e-200)
+
 # Physical constants in CGS
+Msun = const.M_sun.cgs.value
 cLight = const.c.cgs.value
 G = const.G.cgs.value
 eCharge = const.e.gauss.value
@@ -21,7 +23,5 @@ nuConst = eCharge / (twopi * me * cLight)
 PmbConst = twopi * np.sqrt(3) * eCharge**2 / cLight
 jmbConst = np.sqrt(3) * eCharge**2 / (2. * cLight)
 ambConst = np.sqrt(3) * eCharge**2 / (4. * me * cLight)
-chunche_c100g100 = 2.2619939050180366385e-6
-chunche_c100g20 = 2.1157699720918349273e-1
 
-Msun = const.M_sun.cgs.value
+
