+++ pysrc/misc_func.py
@@ -1,0 +1,21 @@
+
+
+def clamp(n, min, max):
+    if (n < min):
+        return min
+    elif (n > max):
+        return max
+    else:
+        return n
+
+def clamp_arr(n, min, max):
+    for i in range(len(n)):
+        if (n[i] < min):
+            n[i] = min
+        else:
+            n[i] = max
+
+    return n
+
+
+
